For achievement manager use preconditions separately. Save manager in another folder.
