+++ src/engine/core/managers/events/AbstractTimersManager.ts
@@ -1,7 +1,7 @@
 import { time_global } from "xray16";
 
 import { AbstractManager } from "@/engine/core/managers/base/AbstractManager";
-import { IIntervalDescriptor, ITimeoutDescriptor } from "@/engine/core/managers/events/types";
+import { IIntervalDescriptor, ITimeoutDescriptor } from "@/engine/core/managers/events/events_types";
 import { assert } from "@/engine/core/utils/assertion";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import { getTableSize } from "@/engine/core/utils/table";
