+++ src/engine/core/managers/events/EventsManager.ts
@@ -1,5 +1,5 @@
 import { AbstractTimersManager } from "@/engine/core/managers/events/AbstractTimersManager";
-import { EGameEvent } from "@/engine/core/managers/events/types";
+import { EGameEvent } from "@/engine/core/managers/events/events_types";
 import { assert } from "@/engine/core/utils/assertion";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import { AnyArgs, AnyCallable, AnyContextualCallable, AnyObject, Optional, TCount } from "@/engine/lib/types";
