+++ src/engine/core/managers/events/index.ts
@@ -1,2 +1,2 @@
 export * from "@/engine/core/managers/events/EventsManager";
-export * from "@/engine/core/managers/events/types";
+export * from "@/engine/core/managers/events/events_types";
