--- src/engine/core/managers/index.ts
@@ -1,1 +1,0 @@
-export * from "@/engine/core/managers/base/AbstractManager";
