+++ src/engine/core/managers/interaction/achievements/index.ts
@@ -1,2 +1,2 @@
 export * from "@/engine/core/managers/interaction/achievements/AchievementsManager";
-export * from "@/engine/core/managers/interaction/achievements/types";
+export * from "@/engine/core/managers/interaction/achievements/achievements_types";
