+++ src/engine/scripts/declarations/callbacks/game.ts
@@ -1,5 +1,5 @@
-import { SaveManager } from "@/engine/core/managers/base/SaveManager";
 import { TradeManager } from "@/engine/core/managers/interaction/trade/TradeManager";
+import { SaveManager } from "@/engine/core/managers/save/SaveManager";
 import { smartCoversList } from "@/engine/core/objects/animation/smart_covers";
 import { GameOutroManager } from "@/engine/core/ui/game/GameOutroManager";
 import { extern } from "@/engine/core/utils/binding";
