+++ src/engine/scripts/register/managers_registrator.test.ts
@@ -1,7 +1,6 @@
 import { beforeEach, describe, expect, it } from "@jest/globals";
 
 import { registry } from "@/engine/core/database";
-import { SaveManager } from "@/engine/core/managers/base/SaveManager";
 import { ProfilingManager } from "@/engine/core/managers/debug/ProfilingManager";
 import { EventsManager } from "@/engine/core/managers/events";
 import { AchievementsManager } from "@/engine/core/managers/interaction/achievements";
@@ -17,6 +16,7 @@ import { ItemUpgradesManager } from "@/engine/core/managers/interface/ItemUpgrad
 import { LoadScreenManager } from "@/engine/core/managers/interface/LoadScreenManager";
 import { PdaManager } from "@/engine/core/managers/interface/PdaManager";
 import { StatisticsManager } from "@/engine/core/managers/interface/statistics";
+import { SaveManager } from "@/engine/core/managers/save/SaveManager";
 import { SimulationBoardManager } from "@/engine/core/managers/simulation/SimulationBoardManager";
 import { DynamicMusicManager } from "@/engine/core/managers/sounds/DynamicMusicManager";
 import { GlobalSoundManager } from "@/engine/core/managers/sounds/GlobalSoundManager";
